feat: show IPQ score on simulation dashboard
