+++ frontend/src/pages/Dashboard.jsx
@@ -82,6 +82,7 @@ export default function AdminDashboard() {
                             <th className="px-4 py-3">Audio Encode (ms)</th>
                             <th className="px-4 py-3">Transcript (ms)</th>
                             <th className="px-4 py-3">TFFT (ms)</th>
+                            <th className="px-4 py-3">IPQ Score</th>
                             <th className="px-4 py-3">Created</th>
                         </tr>
                     </thead>
@@ -102,6 +103,7 @@ export default function AdminDashboard() {
                                 <td className="px-4 py-3">{s.audio_encode_ms}</td>
                                 <td className="px-4 py-3">{s.transcript_ms}</td>
                                 <td className="px-4 py-3">{s.tfft_ms}</td>
+                                <td className="px-4 py-3">{s.ipq_score ?? "N/A"}</td>
                                 <td className="px-4 py-3">
                                     {new Date(s.created_at).toLocaleString()}
                                 </td>
